feat(docs): enable local docs preview for react wrapper packages (#118)
Decouple docs server from the gulp tasks and extract it in a separate file.
Honour default exports when bootstrapping modules.
Include legacy kendo configuration only when kendo is imported.
